+++ types/index.d.ts
@@ -1,3 +1,4 @@
 export { default } from "./src/wallet-manager-evm-hinkal.js";
 export { default as WalletAccountEvmHinkal } from "./src/wallet-account-evm-hinkal.js";
-export { HinkalError, ProviderNotConnectedError, InvalidRecipientError, InvalidAmountError } from "./src/errors.js";
+export { HinkalError, InvalidRecipientError, InvalidAmountError } from "./src/errors.js";
+export { WdkError, InvalidSignerError, ProviderRequiredError } from "@tetherto/wdk-wallet";